Modified the publish script to commit output to blog repo output and the source to origin

+++ publish.sh
@@ -1,7 +1,20 @@
 #!/bin/bash
+pelican content/ > /dev/null
+if [ "$?" -ne "0" ];then
+  echo "Pelican is not installed or did not convert /content to /output"
+  exit
+fi
 cd output
 git init
 git remote add output https://github.com/tsamis73/tsamis73.github.io.git
 git add *
 git commit -a 
 git push output master -f
+if [ "$?" -e "0" ];then                                                                                                                                                                                           
+  echo "Push to output ompleted"
+fi
+cd ..
+git add *
+git commit -a
+git push origin master
+
